More Swift Linting
